feat: small bug fixes
